connlib: limit the number of relays used per-connection (#2514)
Fixes #2499 

@AndrewDryga This should be accompanied by a portal-side PR that:
* Sorts gateway somehow
* Stop sending STUN servers: we only need TURN URLs since webrtc already
fall back onto STUN URLs when it fails to allocate.

Right now, if we merge this without removing TURN URLs we might not get
a connection where TURN is needed.

I think we should leave the `Stun` type in the message, for hole-punched
only connections that we want to implement in the function. Which from
connlib's side shouldn't need any change if we lave it like this, the
portal should only send STUN urls in that case.

---------

Signed-off-by: Gabi <gabrielalejandro7@gmail.com>
Co-authored-by: Andrew Dryga <andrew@dryga.com>