Consume device upon `add_route`
At the moment, we use an async lock to hold it across an await point. This is very prone to dead-locks and overall does not play well with `poll`-style code.

Instead of holding the lock, we simply remove the device from the `Option`, making it inaccessible to all other components. This doesn't _block_ other code but also enforces a consistent state.