Distribute apps on version bumps and manual trigger instead of PR labels (#2535)
Unfortunately the label-based method of distributing app builds was
flawed because it was trigger on `push: main` which doesn't include the
PR label metadata in the event.

Instead, we can (and probably should) distribute new app builds to
testers whenever the patch version changes. This PR aims to do that.

What this means:
- We'll want to bump the patch version to distribute new app builds to
Test Flight and Play Console. I think this is ok, because once the build
goes to Test Flight or Play Console for distribution, it should be
potentially publishable. This is the assumption both Google and Apple
make throughout their publishing workflow (and why we got temporarily
banned by AppStore Connect).
- Day to day development and testing requires building locally, which I
think is OK since it's relatively easy to do so.