Use `ArcSwapOption` for `device`
An `ArcSwapOption` is a container for an `Arc` that can be changed atomically.

In order to make use of this we wrap the `ArcSwapOption` in _another_ `Arc`. This allows us to share a _pointer_ to this atomic storage with all `peer_handlers`. Thus, each peer handler can always load the latest device at any time and we can atomically swap out the device whilst still sharing the _same_ device across all usages, i.e. the "inner" `Arc` of the `ArcSwapOption`.