On unexpected tunnel shutdown show an alert
The alert is shown from the tunnel side because on iOS, the tunnel could be on
while the app is not in the foreground.

We do the same for macOS for consistency.

However, to show an alert from the tunnel side, we are
forced to use a deprecated API (`displayMessage()`).