connlib: no longer limit the number of relays to use
We now expect the portal to limit those. As we currently did this could
coincidentally, prevent us from using ipv6/ipv4 since we might pick only
ipv4/ipv6 addresses for them.

I believe this means we will use 4 addresess instead of 2, this makes
connection startup time a bit slower. However, if we used a SRV address
it would be the same and normally in webrtc you use 2 SRV addresses,
which resolve(in some cases) to 4 ips.