Shutdown tunnel on quit in macos (#2710)
Fixes #2565

With this PR, the 'Quit' menu item (in the macOS status item menu)
disconnects the tunnel (if tunnel is active) before quitting. While the
tunnel is connected, the menu item's title is changed to 'Disconnect and
Quit'.

Depends on #2687 (and this PR currently includes those commits), so
opened as draft.

**Edit:** Now made ready for review.