Do retry-then-signout when startTunnel() fails as well (#2723)
Fixes #2718

It appears that #2687 missed the case where the tunnel failed to start
for some reason, so the tunnel status never goes to "Connecting" and
then back to "Disconnected" -- we were reacting to the tunnel status
becoming "Disconnected".

In this PR, we do the retry when startTunnel() throws an error as well.