Add a log tab in settings (#2895)
Fixes #2725.

This PR adds a new tab in settings called "Diagnostic Logs". It has:

  - A label saying: "Log directory size: `<size>`"
  - A button to clear the logs
  - A button to export logs

We calculate the log directory size when the user navigates to that tab.
When “Clear logs” button is clicked, we can stop the calculation (if
it’s going on), and then clear the logs. We refresh the log size every 1
min. We stop refreshing when the user goes to another tab. We’ll refresh
when the user navigates to the Log tab.