fix(windows): add routes faster (#3118)
This will fix #3114 and save about 13 seconds at startup, compared to
shelling out to Powershell.

I'm not 100% sure it works for IPv6 routes - I'm setting port, flowinfo,
and scope to 0 and just assuming that it's fine.

For some reason Windows wants a socket address in this API even though I
don't think the port is used for anything.