fix(Apple): Attempt to fix window selectedness of the authentication webview (#3126)
Attempt to fix #2881.

I can't reproduce the exact issue anymore, but I'm guessing activating
the app causes the web view window to lose selectedness. So we don't do
that in the PR.

Also, this PR fixes the scenario where the app is quit while the web
view is shown -- we now close the webview window in that case.