fix(gateway): separete ipv4 and ipv6 queries for dns resolving
Some dns servers return NXDOMAIN for queries where the address exists but there is no
answer for the given query type(e.g. AAAA-only records). This is not up to spec and
musl PROPERLY assumes that means there is no record of any type. Saddly, this happens even
with google DNS so we can expect it to happen everywhere. So we use getaddrinfo to separate
requests for A and AAAA queries and preventing this.