feat(windows): crash handler that writes minidumps to ... (#3200)
... ProgramData/dev.firezone.client/dumps/last_crash.dmp, in debug
builds only for now

For #3111.
I was way off on my estimated dump size. It's only about 500 KB.

Please double-check that the Linux exes are still stripped properly.
Windows, Linux, and Mac have different ways to do embedded or
non-embedded debug info, so the Windows client should still be 20 MB but
with a ~62MB pdb next to it, and the Linux client should still be 15 MB.

- [ ] Check Linux artifacts weren't accidentally affected
- [ ] Make sure connlib's panic handler doesn't interfere with this
- [ ] Later on, change CI so pdb files are saved as artifacts
- [ ] Later on, enable for release builds

---------

Signed-off-by: Reactor Scram <ReactorScram@users.noreply.github.com>