feat(windows): prototype module for IPC (#3304)
To work around #2975 I want to run connlib in a subprocess.
We're already using Tokio' `named_pipe` module for deep links, so I made
a generic request-response IPC system with it.

This can get merged in as its own PR, and the actual moving of the
connlib `Session` will be a later PR atop this.