feat(connection): introduce keep-alive and expose `last_seen` (#3388)
We configure wireguard's keep-alive to 5 seconds and patch `boringtun`
to expose the time since the last packet, which will always be <
KEEP_ALIVE (5 seconds) if the connection is intact. The policy on what
to do on missed keep-alives is pushed to the upper layer. Instead of
acting on it, we simply expose a `stats` function that exposes the data.

An upper layer can then decide on what to do in the case on missed
keep-alives.

Resolves: #3372.