refactor(windows): set absolute paths for logs and wintun.dll (#3428)
This is part of fixing #3425 

Until now I changed the app's working directory into our %LOCALAPPDATA%
folder and then used relative paths.

But this causes two problems:
- Passing `.\wintun.dll` when loading the DLL can cause Windows to
search for the DLL. We don't want it to search, we want to put the DLL
in one place and make sure it uses that, since that's the version we'll
be updating
- It means I've been using the app's current working dir as de-facto
global mutable state

The reason I could not fix it sooner is that it needed the bundle ID to
be available before Tauri starts.

---------

Signed-off-by: Reactor Scram <ReactorScram@users.noreply.github.com>
Co-authored-by: Gabi <gabrielalejandro7@gmail.com>