fix(connection): only emit `Transmit.src` that correspond to local sockets (#3411)
It turns out that we need to do some post-processing of the
`Transmit.source` attribute from `str0m`. In its current state, `str0m`
may also set that to a server-reflexive address which is **not** a local
socket. There is a longer discussion around this here:
https://github.com/algesten/str0m/issues/453.

This depends on an unmerged PR in `str0m`:
https://github.com/algesten/str0m/pull/455.