fix(snownet): always clear state of failed connections (#3581)
Initially, I thought it could be useful to keep the connection around in
case we want to do an ICE restart. We don't do that (yet) so for now,
the safer option is to just clean up all the state and let the upper
layers deal with reestablishing it.

This should hopefully fix the repeated emitting of the
`ConnectionFailed` event.